# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: KODE Printer API
  version: 1.0.0
  description: |
    Central print, scan and audit for KODE Sports Club.

    Every non-2xx body is the same envelope (`Error` below) and carries a
    `requestId` that also appears in the `X-Request-Id` response header — it is
    the join key between a user's complaint and the logs.

    Lists are keyset-paginated: pass the `nextCursor` from the previous page as
    `cursor`. Offset pagination is not offered at any endpoint.
  license:
    name: Internal, confidential — KODE Sports Club · Technology
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /printers:
    get:
      tags: [printers]
      summary: List printers
      description: Scoped to the caller's grants. Administrators see the whole fleet.
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: zoneId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string, enum: [online, offline, degraded, unknown] } }
        - { name: search, in: query, schema: { type: string } }
        - { name: includeInactive, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: A page of printers.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrinterPage' }
    post:
      tags: [printers]
      summary: Add a printer by IP address
      description: |
        Serial, vendor, model and capabilities are discovered by probe rather
        than typed. The address must be a private IPv4, and so must the host of
        any `ippUri` supplied with it.
      responses:
        '201':
          description: Created; the probe runs out of band.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Printer' }
        '409': { $ref: '#/components/responses/Error' }
  /printers/{id}:
    get:
      tags: [printers]
      summary: One printer
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The printer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Printer' }
        '403': { $ref: '#/components/responses/Error' }
    put:
      tags: [printers]
      summary: Update a printer
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Updated.
    delete:
      tags: [printers]
      summary: Delete a printer
      description: Refused once the printer has job history — disable it instead, which keeps the record readable.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204':
          description: Deleted.
        '409': { $ref: '#/components/responses/Error' }
  /printers/{id}/active:
    put:
      tags: [printers]
      summary: Enable or disable a printer
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Updated.
  /printers/{id}/drain:
    put:
      tags: [printers]
      summary: Maintenance mode
      description: Finish what is queued, accept nothing new. Distinct from disabling.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Updated.
  /printers/{id}/probe:
    post:
      tags: [printers]
      summary: Re-run the capability probe
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The printer, with capabilities refreshed.
  /printers/{id}/check-status:
    post:
      tags: [printers]
      summary: Diagnostics
      description: Port reachability on 631 and 9100, plus whether walk-up tracking is available at all.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Diagnostic report.
  /printers/{id}/qr:
    get:
      tags: [printers]
      summary: QR sticker for this printer
      description: Scanning it opens the print page with this printer selected.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: An SVG.
          content:
            image/svg+xml:
              schema: { type: string }
  /printers/{id}/print-file:
    post:
      tags: [printers]
      summary: Submit a print job
      description: |
        Returns 202: conversion and dispatch happen out of band and the outcome
        arrives over `/stream`. A job above the warning threshold is refused once
        with `requiresConfirmation` in the details; resubmit with
        `confirmLargeJob` to proceed.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                options: { type: string, description: JSON-encoded print options. }
                confirmLargeJob: { type: boolean }
      responses:
        '202':
          description: Queued, or held for release at the device.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrintAccepted' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
...
  /templates/{id}/print:
    post:
      tags: [templates]
      summary: Print a template
      description: Goes through the ordinary pipeline, so the safety gate, the ledger entry and the audit row all still happen.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '202':
          description: Queued.
...
  /users/{id}/printers:
    put:
      tags: [users]
      summary: Replace an account's printer grants
      description: The only write path to printer permissions. The whole set is replaced, so the audit row holds before and after.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Updated.
...
  /collectors/agent/printers:
    get:
      tags: [collectors]
      summary: The printer roster for this collector
      description: The one response that carries SNMP credentials, and the reason it is sent with Cache-Control no-store.
      security: [{ collectorKey: [] }]
      responses:
        '200':
          description: Printers assigned to this collector.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    collectorKey:
      type: http
      scheme: bearer
      description: A per-collector API key. Never valid on an administrative route.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    Cursor:
      name: cursor
      in: query
      description: Opaque keyset cursor from the previous page.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    Error:
      description: The standard error envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              description: Machine-readable code from the error taxonomy.
            message:
              type: string
              description: Written for the person reading it, not for a log.
            details:
              type: object
              additionalProperties: true
              description: Never contains a stack trace, a SQL string or a secret.
            requestId:
              type: string
    Page:
      type: object
      required: [items, hasMore, nextCursor]
      properties:
        items:
          type: array
          items: {}
        hasMore: { type: boolean }
        nextCursor:
          type: [string, 'null']
    Printer:
      type: object
      description: |
        Carries no credentials. snmpConfigured reports whether one exists; the
        value itself has no representation on this shape at all.
      properties:
        id: { type: integer }
        name: { type: string }
        ipAddress: { type: string }
        area: { type: [string, 'null'] }
        zoneId: { type: [integer, 'null'] }
        zoneLabel: { type: [string, 'null'] }
        vendor: { type: [string, 'null'] }
        model: { type: [string, 'null'] }
        status: { type: string, enum: [online, offline, degraded, unknown] }
        stateReasons:
          type: array
          items: { type: string }
          description: IPP-style keywords such as media-empty, media-jam, device-mismatch.
        transport: { type: string, enum: [auto, ipp, raw9100] }
        capabilities:
          type: object
          additionalProperties: true
          description: probedVia none means we asked and got no answer. Present those options as unverified rather than implying they take effect.
        snmpConfigured: { type: boolean }
        supplies:
          type: array
          items: { $ref: '#/components/schemas/PrinterSupply' }
        isActive: { type: boolean }
        isDraining: { type: boolean }
        maxJobImpressions: { type: [integer, 'null'] }
        walkupTrackingUnavailable:
          type: boolean
          description: True when this device reports no walk-up activity at all. Every report covering it must say so.
    PrinterSupply:
      type: object
      properties:
        name: { type: string }
        colorant: { type: [string, 'null'] }
        percent: { type: [number, 'null'] }
        estimatedDaysRemaining:
          type: [integer, 'null']
          description: Withheld until the history supports it, rather than extrapolated from two readings.
    PrinterPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Printer' }
...
    PrintAccepted:
      type: object
      properties:
        jobId: { type: integer }
        status: { type: string, enum: [queued, held] }
        impressions: { type: integer }
        message: { type: string }
...